# swagger.yaml
openapi: 3.0.3
info:
  title: Troud API
  description: |
    **Troud** es una aplicación backend tipo "Tinder" para intercambio de ropa que permite a los usuarios buscar, encontrar e intercambiar prendas de vestir.
    
    ## Características Principales
    - Gestión de usuarios con autenticación JWT
    - CRUD completo de artículos de ropa
    - Sistema de reacciones (Like, SuperLike, DisLike)
    - Sistema de matches automático
    - Subida de imágenes a AWS S3
    
    ## Autenticación
    La API utiliza dos tipos de autenticación:
    - **Basic Auth**: Para obtener el token inicial (username=phone, password=password)
    - **JWT Bearer**: Para endpoints protegidos (Header: `Authorization: Bearer <token>`)
  version: 1.0.0
  contact:
    name: Troud Team
    url: https://github.com/personal/troud
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo local
  - url: https://troud.vercel.app/api
    description: Servidor de producción
...
paths:
...
  /articles:
    get:
      tags:
        - Articles
      summary: Listar artículos
      security:
        - BearerAuth: []
      parameters:
        - name: phoneOwner
          in: query
          description: Filtrar por teléfono del propietario
          required: false
          schema:
            type: string
            pattern: '^[0-9]{7,10}$'
      responses:
        '200':
          description: Lista de artículos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  message:
                    type: string
                    example: "articles listed"
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      tags:
        - Articles
      summary: Crear nuevo artículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleRequest'
      responses:
        '201':
          description: Artículo creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: ID del artículo creado
                  message:
                    type: string
                    example: "Article created"
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /articles/{articleId}:
    get:
      tags:
        - Articles
      summary: Obtener artículo por ID
      parameters:
        - name: articleId
          in: path
          required: true
          description: ID del artículo
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Artículo encontrado con estadísticas de reacciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Article'
                  like:
                    type: integer
                    description: Número de likes
                  superLikes:
                    type: integer
                    description: Número de superlikes
                  dislikes:
                    type: integer
                    description: Número de dislikes
                  message:
                    type: string
                    example: "article retrieved"
    
    put:
      tags:
        - Articles
      summary: Actualizar artículo
      security:
        - BearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          description: ID del artículo
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleRequest'
      responses:
        '200':
          description: Artículo actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: ID del artículo actualizado
                  message:
                    type: string
                    example: "Article updated"
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    delete:
      tags:
        - Articles
      summary: Eliminar artículo
      parameters:
        - name: articleId
          in: path
          required: true
          description: ID del artículo
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Artículo eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: ID del artículo eliminado
                  message:
                    type: string
                    example: "Article deleted"
  /articles/categories/{category}/{phoneUser}:
    get:
      tags:
        - Articles
      summary: Obtener artículos por categoría
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          description: Categoría del artículo
          schema:
            type: string
            maxLength: 60
        - name: phoneUser
          in: path
          required: true
          description: Teléfono del usuario
          schema:
            type: string
            pattern: '^[0-9]{7,10}$'
      responses:
        '200':
          description: Lista de artículos por categoría
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  message:
                    type: string
                    example: "articles listed"
        '401':
          $ref: '#/components/responses/Unauthorized'
  /articles/unreaction/{phoneUser}:
    get:
      tags:
        - Articles
      summary: Obtener artículos sin reacción
      description: Obtiene artículos que el usuario no ha reaccionado
      security:
        - BearerAuth: []
      parameters:
        - name: phoneUser
          in: path
          required: true
          description: Teléfono del usuario
          schema:
            type: string
            pattern: '^[0-9]{7,10}$'
      responses:
        '200':
          description: Lista de artículos sin reacción
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  message:
                    type: string
                    example: "article without reaction retrieved"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /images:
    get:
      tags:
        - Images
      summary: Listar imágenes
      responses:
        '200':
          description: Lista de imágenes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Image'
                  message:
                    type: string
                    example: "Images listed"
    
    post:
      tags:
        - Images
      summary: Subir nueva imagen
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: Archivo de imagen a subir
      responses:
        '201':
          description: Imagen subida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    description: ID de la imagen creada
                  path:
                    type: object
                    properties:
                      profilePicture:
                        type: string
                        description: URL de la imagen subida
                  message:
                    type: string
                    example: "image created"
        '400':
          description: Error al subir la imagen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  errors:
                    type: object
                    properties:
                      title:
                        type: string
                        example: "Image Upload Error"
                      detail:
                        type: string
                        description: Descripción del error
                      error:
                        type: object
                        description: Objeto de error detallado
  /images/{imageId}:
    get:
      tags:
        - Images
      summary: Obtener imagen por ID
      parameters:
        - name: imageId
          in: path
          required: true
          description: ID de la imagen
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Imagen encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Image'
                  message:
                    type: string
                    example: "image retrieved"
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Autenticación básica usando teléfono como username y contraseña
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/token
  responses:
    Unauthorized:
      description: Token de acceso faltante o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Article:
      type: object
      properties:
        _id:
          type: string
          description: ID único del artículo
          example: "507f1f77bcf86cd799439013"
        phoneOwner:
          type: string
          pattern: '^[0-9]{7,10}$'
          description: Teléfono del propietario
          example: "3001234567"
        idOwner:
          type: string
          description: ID del propietario
          example: "507f1f77bcf86cd799439011"
        type:
          type: string
          maxLength: 60
          description: Tipo de prenda
          example: "Camiseta"
        size:
          type: string
          maxLength: 10
          description: Talla
          example: "M"
        name:
          type: string
          maxLength: 60
          description: Nombre del artículo
          example: "Camiseta Nike Deportiva"
        brand:
          type: string
          maxLength: 60
          description: Marca
          example: "Nike"
        condition:
          type: string
          maxLength: 15
          description: Estado de la prenda
          example: "Usado - Buen estado"
        gender:
          type: string
          maxLength: 10
          description: Género
          example: "Unisex"
        description:
          type: string
          maxLength: 120
          description: Descripción detallada
          example: "Camiseta deportiva en excelente estado, ideal para ejercicio"
        color:
          type: string
          maxLength: 30
          description: Color principal
          example: "Azul"
        date:
          type: string
          description: Fecha de creación
          example: "12/15/2023 10:30:25 AM"
        urlPhoto:
          type: string
          description: URL de la foto del artículo
          example: "https://s3.amazonaws.com/bucket/article.jpg"
        city:
          type: string
          maxLength: 60
          description: Ciudad donde se encuentra
          example: "Bogotá"
    CreateArticleRequest:
      type: object
      required:
        - phoneOwner
        - idOwner
        - type
        - size
        - name
        - brand
        - condition
        - gender
        - description
        - color
        - urlPhoto
        - city
      properties:
        phoneOwner:
          type: string
          pattern: '^[0-9]{7,10}$'
          description: Teléfono del propietario
          example: "3001234567"
        idOwner:
          type: string
          description: ID del propietario
          example: "507f1f77bcf86cd799439011"
        type:
          type: string
          maxLength: 60
          description: Tipo de prenda
          example: "Camiseta"
        size:
          type: string
          maxLength: 10
          description: Talla
          example: "M"
        name:
          type: string
          maxLength: 60
          description: Nombre del artículo
          example: "Camiseta Nike Deportiva"
        brand:
          type: string
          maxLength: 60
          description: Marca
          example: "Nike"
        condition:
          type: string
          maxLength: 15
          description: Estado de la prenda
          example: "Usado - Buen estado"
        gender:
          type: string
          maxLength: 10
          description: Género
          example: "Unisex"
        description:
          type: string
          maxLength: 120
          description: Descripción detallada
          example: "Camiseta deportiva en excelente estado, ideal para ejercicio"
        color:
          type: string
          maxLength: 30
          description: Color principal
          example: "Azul"
        urlPhoto:
          type: string
          description: URL de la foto del artículo
          example: "https://s3.amazonaws.com/bucket/article.jpg"
        city:
          type: string
          maxLength: 60
          description: Ciudad donde se encuentra
          example: "Bogotá"
    UpdateArticleRequest:
      type: object
      properties:
        phoneOwner:
          type: string
          pattern: '^[0-9]{7,10}$'
          description: Teléfono del propietario
          example: "3001234567"
        idOwner:
          type: string
          description: ID del propietario
          example: "507f1f77bcf86cd799439011"
        type:
          type: string
          maxLength: 60
          description: Tipo de prenda
          example: "Camiseta"
        size:
          type: string
          maxLength: 10
          description: Talla
          example: "M"
        name:
          type: string
          maxLength: 60
          description: Nombre del artículo
          example: "Camiseta Nike Deportiva"
        brand:
          type: string
          maxLength: 60
          description: Marca
          example: "Nike"
        condition:
          type: string
          maxLength: 15
          description: Estado de la prenda
          example: "Usado - Buen estado"
        gender:
          type: string
          maxLength: 10
          description: Género
          example: "Unisex"
        description:
          type: string
          maxLength: 120
          description: Descripción detallada
          example: "Camiseta deportiva en excelente estado, ideal para ejercicio"
        color:
          type: string
          maxLength: 30
          description: Color principal
          example: "Azul"
        date:
          type: string
          description: Fecha de actualización
          example: "12/15/2023 10:30:25 AM"
        urlPhoto:
          type: string
          description: URL de la foto del artículo
          example: "https://s3.amazonaws.com/bucket/article.jpg"
        city:
          type: string
          maxLength: 60
          description: Ciudad donde se encuentra
          example: "Bogotá"
...
    Image:
      type: object
      properties:
        _id:
          type: string
          description: ID único de la imagen
          example: "507f1f77bcf86cd799439016"
        profilePicture:
          type: string
          description: URL de la imagen almacenada en S3
          example: "https://s3.amazonaws.com/bucket/image.jpg"
    Error:
      type: object
      properties:
        data:
          type: string
          nullable: true
          example: null
        message:
          type: string
          description: Mensaje de error descriptivo
          example: "Validation error" 